v1.0 (beta)

+++ app/src/main/java/com/myfablo/seller/manage/orders/OrderHistoryRecyclerAdapter.java
@@ -72,7 +72,7 @@ public class OrderHistoryRecyclerAdapter extends RecyclerView.Adapter<OrderHisto
         }
     }
 
-    private String getPriceWithSymbol(int amount) {
+    private String getPriceWithSymbol(Float amount) {
         return "₹" + amount;
     }
