+++ app/src/main/java/com/myfablo/seller/orders/OrderRecyclerAdapter.java
@@ -143,7 +143,7 @@ public class OrderRecyclerAdapter extends RecyclerView.Adapter<OrderRecyclerAdap
         }
     }
 
-    private String getPriceWithSymbol(int amount) {
+    private String getPriceWithSymbol(Float amount) {
         return "₹" + amount;
     }
